+++ .betterer.results
@@ -1,5 +1,5 @@
 // BETTERER RESULTS V2.
-//
+// 
 // If this file contains merge conflicts, use `betterer merge` to automatically resolve them:
 // https://phenomnomnominal.github.io/betterer/docs/results-file/#merge
 //
@@ -1185,7 +1185,7 @@ exports[`better eslint`] = {
       [0, 0, 0, "Unexpected any. Specify a different type.", "5"],
       [0, 0, 0, "Unexpected any. Specify a different type.", "6"]
     ],
-    "public/app/core/components/AppChrome/DockedMegaMenu/NavFeatureHighlight.tsx:5381": [
+    "public/app/core/components/AppChrome/DockedMegaMenu/FeatureHighlight.tsx:5381": [
       [0, 0, 0, "Styles should be written using objects.", "0"]
     ],
     "public/app/core/components/AppChrome/MegaMenu/NavFeatureHighlight.tsx:5381": [
